♻️ [Refactoring]: #221 application レイヤーを純粋なユースケース層に整理
spec 046 のステップ 2〜5 + リネームを一括実施し、`src/application/`
をユースケース関数のみを持つレイヤーに再編する。

- 値オブジェクト移動 (PR 2)
  - `TargetId` → `src/target/id.rs`
  - `ScopedPath` / `normalize_path` / `resolve_nonexistent_path` → `src/component/scoped_path.rs`
  - `FileOperation` → `src/component/file_operation.rs`
  - `src/application/plugin_action_types.rs` を削除
- ドメイン要素昇格 (PR 3)
  - `PluginAction` → `src/plugin/action.rs`
  - `PluginIntent` / `ExpandResult` → `src/plugin/intent.rs`
  - `PluginDeployment` + `load_plugin_deployment` / `cleanup_plugin_directories` → `src/plugin/deployment.rs`
  - `src/application/plugin_{action,intent,deployment}.rs` を削除
- ユースケース関数整理 (PR 4)
  - `list_installed` → `src/plugin/cache.rs`（`pub(crate)`）
  - `list_all_placed` → `src/target/placed.rs`（`pub(crate)`）
- ファイル名リネーム (PR 5)
  - `plugin_info.rs` → `info.rs`
  - `plugin_catalog.rs` → `catalog.rs`
  - `plugin_operations.rs` → `lifecycle.rs`
  - `src/application.rs` の `mod` / `pub use` を新名に更新

Refs #221
