+++ src/application/catalog_test.rs
