+++ src/application/info_test.rs
