+++ src/application/lifecycle_test.rs
