+++ src/component/scoped_path.rs
@@ -1,35 +1,7 @@
-//! プラグインアクション型定義
-//!
-//! ターゲット識別子、スコープ付きパス、ファイル操作の値オブジェクト群。
+//! スコープ付きパス（project_root 配下保証の値オブジェクト）
 
 use crate::error::{PlmError, Result};
 use std::path::{Path, PathBuf};
-
-/// ターゲット識別子（型安全）
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct TargetId(String);
-
-impl TargetId {
-    pub fn new(name: impl Into<String>) -> Self {
-        Self(name.into())
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl From<&str> for TargetId {
-    fn from(s: &str) -> Self {
-        Self::new(s)
-    }
-}
-
-impl std::fmt::Display for TargetId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
 
 /// スコープ付きパス（project_root配下であることを保証）
 ///
@@ -175,39 +147,6 @@ fn resolve_nonexistent_path(path: &Path) -> Result<PathBuf> {
     Ok(normalize_path(path))
 }
 
-/// 低レベルファイル操作（内部用）
-#[derive(Debug, Clone)]
-pub enum FileOperation {
-    CopyFile { source: PathBuf, target: ScopedPath },
-    CopyDir { source: PathBuf, target: ScopedPath },
-    RemoveFile { path: ScopedPath },
-    RemoveDir { path: ScopedPath },
-}
-
-impl FileOperation {
-    /// 操作の種類を文字列で取得
-    pub fn kind(&self) -> &'static str {
-        match self {
-            FileOperation::CopyFile { .. } => "copy_file",
-            FileOperation::CopyDir { .. } => "copy_dir",
-            FileOperation::RemoveFile { .. } => "remove_file",
-            FileOperation::RemoveDir { .. } => "remove_dir",
-        }
-    }
-
-    /// ターゲットパスを取得
-    pub fn target_path(&self) -> &Path {
-        match self {
-            FileOperation::CopyFile { target, .. } | FileOperation::CopyDir { target, .. } => {
-                target.as_path()
-            }
-            FileOperation::RemoveFile { path } | FileOperation::RemoveDir { path } => {
-                path.as_path()
-            }
-        }
-    }
-}
-
 #[cfg(test)]
-#[path = "plugin_action_types_test.rs"]
+#[path = "scoped_path_test.rs"]
 mod tests;
